refactor: prepare shared ACME scheduling infra for a second scheduler
DS TLS ACME enrollment (issue 03) needs its own, independent
CertificateRenewalScheduler instance alongside the existing auth/sign
one. Two prerequisite fixes to the shared lib make that possible
without conflicts or misplaced ownership:

- Extract the HTTP-01 challenge-port Tomcat connector bean into its
  own lib/acme-spring class, AcmeChallengerConfig. Previously it lived
  inside the auth/sign scheduling config; with two independent
  scheduling configs, each declaring that bean would register two
  Tomcat connectors on the same port 80.
- Move AcmeSchedulingConfig out of lib/acme-spring into
  security-server/admin-service's own config package, renamed
  AcmeCertificateRenewalSchedulingConfig - it wires the SS-specific
  AcmeClientWorker (renamed AcmeCertificateRenewalWorker to match), so
  it was misplaced in the shared lib to begin with. The rename also
  anticipates a sibling DsTls...SchedulingConfig for clearer pairing.

CertificateRenewalScheduler itself is untouched. Fixed one stale
cross-reference the move left behind: AcmeConfig's
isAcmeRenewalActive() Javadoc still pointed at the old package path.

Refs: XRDDEV-3290
