+++ src/lib/acme-core/src/main/java/org/niis/xroad/common/acme/config/AcmeConfig.java
@@ -51,7 +51,8 @@ public interface AcmeConfig {
     }
 
     /**
-     * raw usage with default in org.niis.xroad.common.acme.spring.config.AcmeSchedulingConfig.IsAcmeCertRenewalJobsActive
+     * raw usage with default in
+     * org.niis.xroad.securityserver.restapi.config.AcmeCertificateRenewalSchedulingConfig.IsAcmeCertRenewalJobsActive
      */
     boolean isAcmeRenewalActive();
